clippy 안전 lint 9건 정리 (warning 76 → 67)
- unnecessary_cast: TBS_HORZ/TBS_VERT/TBS_BOTH/TBS_NOTICKS, EN_KILLFOCUS,
  EN_CHANGE, (i+1) as usize 의 동일타입 캐스트 제거.
- identity_op: ((255 << 16) | 0) → (255 << 16).
- needless_range_loop: tray::set_sz_tip 의 인덱스 루프를
  iter().take(n).enumerate() 로 교체.

동작 변화 없음 (컴파일러가 타입/값 동일성 보장). cargo check warning 0
유지, cargo clippy warning 76 → 67.

refs: docs/cycle-20260511-2318.md

+++ src/dialogs/file_trans_thread.rs
@@ -212,7 +212,7 @@ fn process_single_file(
 
         // 진행률 업데이트
         *global_current += 1;
-        send_progress_message(progress_hwnd, WM_PROGRESS_UPDATE, (i + 1) as usize, 0);
+        send_progress_message(progress_hwnd, WM_PROGRESS_UPDATE, i + 1, 0);
         send_progress_message(
             progress_hwnd,
             WM_PROGRESS_CURRENT,
